Changed NPC collisions & started fixing the police movement
Changed the way NPC's are collided with, from originally triggering the police script from the PLAYER vehicle, along with the deletion of the NPC on the NPC script.

to both of those functioning all on the NPC script now.

Started fixing the police movement with character controller now as well so that it can properly tail the player vehicle again.