Removed part of the police movement script
Removed the part in which the police tailed the player as it wasn't working with the new charactercontroller system implemented in

It now just simply pops into one location and now has a collider which will destroy NPC's if they come into contact.

Will add more vehicles to make it look like more are tailing you.