Made some changes in the keycard reader
Made a change in the keycard reader minigame in which once the player lets go of the keycard, it will snap back into its original position.

This change was made due to players losing the keycard very often.

Another change was made with the police tailing mechanic in that the law enforcement vehicle will tail right behind the vehicle instead of going into it.